requirements.txt
addict==2.4.0
future==0.18.3
numpy
opencv-python
matplotlib
Pillow==9.4.0
pyyaml
requests
scikit-image==0.19.3
scipy
torch>=1.9
torchvision==0.14.1
timm==0.6.12
einops==0.6.0
flask==2.2.2